-change damage, health, maxhealth and regen to double -add damageMultiplier, getter and setter to CombatEntityModel.java -add upgrades to PlayerModel.java (speed, damage, health and regen) -add upgrades statistics

+++ src/main/java/gui/ingame/entity/HealthBarRenderer.java
@@ -14,7 +14,7 @@ class HealthBarRenderer {
         g.fillRect(0, 0, width, 2);
 
         g.setColor(Color.GREEN);
-        int greenWidth = width * entity.getHealth() / entity.getMaxHealth();
+        int greenWidth = (int) (width * entity.getHealth() / entity.getMaxHealth());
         g.fillRect(0, 0, greenWidth, 2);
     }
 }
